requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiopg==1.4.0
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
et-xmlfile==1.1.0
frozenlist==1.3.3
geographiclib==2.0
geopy==2.3.0
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
typing_extensions==4.5.0
tzdata==2023.3
yarl==1.8.2
